requirements.txt
pytubefix
youtube-transcript-api
pytube
openai-whisper
transformers
torch
spacy
textblob

